+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Future
+- New methods for `Terminal`: `has_input`, `wait_for_input`,
+  `read_ambiguous_timeout`, `read_timeout`, `read_raw`, `read_raw_timeout`,
+  `read_raw_single_timeout` and `read_byte_timeout`.
 - Add texels - drawing images using charactes.
 - Add option for nearest when scaling image.
 - Add new codes.